+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Drop support for TYPO3 7.6 (#248)
 
 ### Fixed
+- Use the new TypoScript file paths in the userfunc tests (#253)
 
 ## 2.2.1